To avoid the @ character being translated to i, add support for setting ALPHA_LATIN1 (aka ISO-8859-1) as MT DataCoding or ZERO (aka GSM7). Thus sending help@tbc.fyi no longer results in helpitbc.fyi.
The issue was that our Infobip main account is set as ISO encoding by default regardless of the MT datacoding parameter, therefore @ character is being correctly interpreted. But our TBC subaccount must be set to accept the encoding we indicate. When we submit messages with datacoding=0, Infobip uses GSM7 encoding and therefore the @ character is not correctly translated.

On the other hand, the Mblox SMPP gateway supports two character encoding systems, GSM 03.38 encoding and Latin 1 (ISO-8859-1) encoding. Latin 1 is the default for our account. Mblox says that to use the default character encoding system for our account, our messages must be sent with a data_coding parameter set to ZERO.

I tested sending the email address to my phone using Mblox, Infobip with ISO default, and Infobip that is set to accept the encoding we indicate.