Added MBLOX_ATLAS as a new providerType to fix the "java.lang.IllegalStateException: MT not found" exceptions on ezr/strata.
The new Mblox platform, which is named Atlas and available at sms3.mblox.com, returns a decimal message Id like "3000002988033283" and specifies a matching decimal message Id in the delivery receipt like "3000002988033283".

The old Mblox platform, which we identify as providerType MBLOX in our SMPP Configs and is available at smpp.psms.us.mblox.com, returns a hexidecimal message Id like "85c5e006" but specifies a corresponding decimal message Id in the delivery receipt like "2244337670".

Since we need to support both platforms, we must remember to set providerType as MBLOX_ATLAS when we connect to sms3.mblox.com and MBLOX when we connect to smpp.psms.us.mblox.com.